+++ MillimanAccessPortal/MillimanAccessPortal/Startup.cs
@@ -127,6 +127,7 @@ namespace MillimanAccessPortal
                     options.Wtrealm = schemeProperties.Wtrealm;
                     options.CallbackPath = $"{options.CallbackPath}-{scheme.Name}";
 
+                    #region WS-Federation middleware event overrides
                     // Event override to add username query parameter to adfs request
                     options.Events.OnRedirectToIdentityProvider = context =>
                     {
@@ -144,6 +145,22 @@ namespace MillimanAccessPortal
                         {
                             context.ProtocolMessage.SetParameter("username", context.Properties.Items["username"]);
                         }
+                        return Task.CompletedTask;
+                    };
+
+                    // Event override to handle all remote failures from WsFederation middleware
+                    options.Events.OnRemoteFailure = context =>
+                    {
+                        context.Response.Redirect("/");
+                        context.HandleResponse();
+                        return Task.CompletedTask;
+                    };
+
+                    // Event override to handle authentication failures from WsFederation middleware
+                    options.Events.OnAuthenticationFailed = context =>
+                    {
+                        context.Response.Redirect("/");
+                        context.HandleResponse();
                         return Task.CompletedTask;
                     };
 
@@ -233,6 +250,7 @@ namespace MillimanAccessPortal
 
                         context.Response.Redirect($"/Account/{nameof(AccountController.ExternalLoginCallback)}");
                     };
+                    #endregion
                 });
             }
             authenticationBuilder.AddIdentityCookies();
@@ -492,7 +510,6 @@ namespace MillimanAccessPortal
 
                 Log.Information("MVC took {elapsed}ms", stopwatch.Elapsed.TotalMilliseconds);
             });
-
 
             app.UseMvc(routes =>
             {