+++ MillimanAccessPortal/MillimanAccessPortal/src/ts/react/login/login-form.tsx
@@ -181,12 +181,11 @@ export class LoginForm extends Form<{}, LoginFormState> {
               this.focusPasswordInput();
             });
           } else {
-            window.location.href = `/Account/RemoteAuthenticate?username=${this.state.data.username}`;
+            window.location.replace(`/Account/RemoteAuthenticate?username=${this.state.data.username}`);
           }
         })
         .catch(() => {
-          errors.username = 'An error occured.';
-          alert(errors.username);
+          errors.username = 'An error occurred.';
           this.setState({
             errors,
             userConfirmed: false,