+++ Documentation/ReleaseNotes.md
@@ -8,6 +8,12 @@
 - Implemented a mechanism for controlling the number of domains that can be added to a client email domain whitelist
 - Redesigned the Forgot Password page
 - Fixed an issue preventing content publication from running in parallel
+- Converted the Forgot Password page to use React
+- Converted the Reset Password page to use React
+
+#### v1.5.1
+
+- Fixed an issue where user login fails in certain unusual circumstances
 
 ### v1.5.0